+++ packages/core/src/common/binding/writable.ts
