+++ packages/core/src/index.ts
@@ -48,7 +48,7 @@ export {BladePlugin, createBladeController} from './blade/plugin';
 export * from './common/binding/ticker/manual';
 export * from './common/binding/ticker/interval';
 export * from './common/binding/ticker/ticker';
-export * from './common/binding/value';
+export * from './common/binding/value/value';
 
 export * from './common/constraint/composite';
 export * from './common/constraint/constraint';
