+++ packages/core/src/input-binding/color/plugin-object.ts
@@ -1,4 +1,4 @@
-import {BindingReader} from '../../common/binding/binding';
+import {BindingReader} from '../../common/binding/writable';
 import {Formatter} from '../../common/converter/formatter';
 import {InputBindingPlugin} from '../plugin';
 import {ColorController} from './controller/color';
