+++ packages/core/src/input-binding/plugin.ts
@@ -2,9 +2,13 @@ import {createBlade} from '../blade/common/model/blade';
 import {InputBindingController} from '../blade/input-binding/controller/input-binding';
 import {LabeledValueController} from '../blade/label/controller/value-label';
 import {LabelPropsObject} from '../blade/label/view/label';
-import {WritableBinding, BindingReader, BindingWriter} from '../common/binding/binding';
+import {
+	BindingReader,
+	BindingWriter,
+	WritableBinding,
+} from '../common/binding/writable';
 import {BindingTarget} from '../common/binding/target';
-import {BindingValue} from '../common/binding/value';
+import {BindingValue} from '../common/binding/value/value';
 import {Constraint} from '../common/constraint/constraint';
 import {Value} from '../common/model/value';
 import {ValueMap} from '../common/model/value-map';
