+++ packages/core/src/monitor-binding/plugin.ts
@@ -1,7 +1,7 @@
 import {createBlade} from '../blade/common/model/blade';
 import {LabelPropsObject} from '../blade/label/view/label';
 import {MonitorBindingController} from '../blade/monitor-binding/controller/monitor-binding';
-import {BindingReader} from '../common/binding/binding';
+import {BindingReader} from '../common/binding/writable';
 import {MonitorBinding} from '../common/binding/monitor';
 import {BindingTarget} from '../common/binding/target';
 import {IntervalTicker} from '../common/binding/ticker/interval';
